deprecated app.py and sam.py fixed
